Langchain/06_image_generation/requirements.txt
streamlit 
openai 
python-dotenv 
requests